requirements.txt
pytest
pytest-asyncio

aioredis
    redis
celery
flowers
aiogran
python-dotenv
alembic
sqlalchemy
asyncpg